- Moved IPipelineBehaviour to be more compatible with MediatR - Minor refactoring
