+++ examples/Flume.Samples/PerformanceBenchmarks.cs
@@ -1,5 +1,4 @@
 using System.Diagnostics;
-using Flume.Pipelines;
 using Microsoft.Extensions.DependencyInjection;
 
 #pragma warning disable CA1303
