+++ src/Flume/Internal/PipelineCompiler.cs
@@ -2,11 +2,6 @@ using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Linq;
-using System.Linq.Expressions;
-using System.Reflection;
-using System.Threading;
-using System.Threading.Tasks;
-using Flume.Pipelines;
 
 namespace Flume.Internal;
